WIP: Add state wrapper and animation to modal
There was no animation of a modal.

To add animation to the modal I decided to add store of modal open
status because buttons which open and close the modal and the modal are
in different places in a app code. Also I decided to remain native open/close
manipulation of modal because If I delete it then I have to write an extra
scss code.

I also added animations of open/close but the close animation don't work
and I don't know why.