Change todo list animations and fix modal closing
When user entered some todo into input of modal and closed it so the
modal added the todo. Also when an user toggle todo status so there were
two same todos. The first that was disapearing and the second that was
appearing.

I change the close modal button type to this button don't add a new todo
if the input has some value. Also I change the list animation to when an
user toggle status of some todo there were not two same todo at one
time.